ride-request-service/requirements.txt
Flask
flask_pymongo
PyJWT
requests
kafka-python